rename twrpdtgen file to twrp_dt_gen to avoid conflicts with package name
