+++ twrpdtgen/__main__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 
-from twrpdtgen.twrpdtgen import main
+from twrpdtgen.twrp_dt_gen import main
 
 if __name__ == '__main__':
     main()
